triact-rag/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
pyjwt
numpy
